# angular/src/swagger/model/projectState.ts
/**
 * HackZurich'18 Vertec 4.0
 * API for the HackZurich'18 Vertec 4.0 Project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ProjectState {
    id?: number;
    projectId?: number;
    actual?: number;
    numberOfTasks?: number;
}