# angular/src/swagger/model/task.ts
/**
 * HackZurich'18 Vertec 4.0
 * API for the HackZurich'18 Vertec 4.0 Project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Task {
    id?: number;
    projectId?: number;
    name?: string;
    description?: string;
    budget?: number;
    startDate?: Date;
    endDate?: Date;
}